Modify Terraform configurations for bfv instances
In the openstack_compute_instance_v2 resource, we now define a block
device as a boot volume of fixed size. And since we need the UUID of the
image the instance uses, we take advantage of the
openstack_images_image_v2 data source. Finally, we introduce two new
variables for setting the image name and the boot volume size.