Deploy booting from an existing volume, define volume lifecycle
Rather than defining the boot volume within the instance's
block_device, define a separate volume resource and use that as the
boot volume.

This way, we can set a lifecycle meta-argument on the boot volume
and configure it to ignore changes to the image UUID.

This means that if we have an existing instance and the image is
cycled so that a new image has the same name, but a different UUID,
Terraform will not touch the boot volume and hence, will leave
the instance alone as well. However, if we destroy the instance and
then respin it, it boots from the new image.